fixed if gloss task delete main task is too

+++ ai_glex/api_views.py
@@ -127,12 +127,14 @@ class GlossaryFileView(viewsets.ViewSet):
         with transaction.atomic():
             if job:
                 objects_to_delete = GlossaryFiles.objects.filter(job=job, id__in=delete_list)
-                terms_ids = objects_to_delete.values_list('id', flat=True).distinct()
-                print("terms_ids--->",terms_ids)
-                ids_str = ','.join(map(str, terms_ids))
-                with connection.cursor() as cursor:
-                    cursor.execute(f"DELETE FROM ai_glex_termsmodel WHERE id IN ({ids_str})")
-                #objects_to_delete.term_file.all().delete()
+                terms_ids =objects_to_delete.term_file.all()
+                if terms_ids:
+                    terms_ids = terms_ids.values_list("id",flat=True).distinct()
+                    print("terms_ids--->",terms_ids)
+                    ids_str = ','.join(map(str, terms_ids))
+                    with connection.cursor() as cursor:
+                        cursor.execute(f"DELETE FROM ai_glex_termsmodel WHERE id IN ({ids_str})")
+                 
                 objects_to_delete.delete()
             else:
                 proj = Project.objects.get(id=project)
@@ -1261,12 +1263,13 @@ def requesting_ner(joined_term_unit):
 import re
 from ai_glex.serializers import CeleryStatusForTermExtractionSerializer
 @task(queue='default')
-def get_ner_with_textunit_merge(file_id,gloss_model_id):
+def get_ner_with_textunit_merge(file_id,gloss_model_id,gloss_job_id):
     # try:    
     file_instance = File.objects.get(id=file_id)
     gloss_model_inst = Glossary.objects.get(id=gloss_model_id)
     file_path = file_instance.get_source_file_path
     path_list = re.split("source/", file_path)
+    gloss_job_ins = Job.objects.get(id=gloss_job_id)
 
     doc_json_path = path_list[0] + "doc_json/" + path_list[1] + ".json"
     with open(doc_json_path,'rb') as fp:
@@ -1288,9 +1291,9 @@ def get_ner_with_textunit_merge(file_id,gloss_model_id):
     file_instance.done_extraction = True
     file_instance.save()
     terms =  list(set(terms))
-    gloss_job_inst = file_instance.gloss_job
+    #gloss_job_inst = file_instance.gloss_job
     #gloss_model_inst = gloss_job_inst.project.glossary
-    termsmodel_instances = [TermsModel(sl_term=term,job=gloss_job_inst,glossary=gloss_model_inst) for term in terms]
+    termsmodel_instances = [TermsModel(sl_term=term,job=gloss_job_ins,glossary=gloss_model_inst) for term in terms]
     TermsModel.objects.bulk_create(termsmodel_instances)
     print("terms_created")
     file_instance.save()
@@ -1325,11 +1328,11 @@ def extraction_text(request):
  
         
         #file_instance.gloss_model = glossary_project
-        file_instance.gloss_job = gloss_job
-        file_instance.save()  # Save term_model
+        #file_instance.gloss_job = gloss_job
+        #file_instance.save()  # Save term_model
         
         celery_instance_ids.append(file_instance.id)
-        celery_id = get_ner_with_textunit_merge.apply_async(args=(file_id,glossary_project.id))
+        celery_id = get_ner_with_textunit_merge.apply_async(args=(file_id,glossary_project.id,gloss_job.id))
         file_instance.celery_id = celery_id
         file_instance.status = "PENDING"
         file_instance.save()  # Save celery status
@@ -1371,24 +1374,25 @@ from django.dispatch import receiver
 @receiver(pre_delete, sender=Glossary)
 def glossary_delete_handler(sender, instance, **kwargs):
     # Perform actions here
+    print("Glossary_id",instance.id)
     print(f'Instance {instance} is being deleted')
 
 @receiver(pre_delete, sender=Project)
 def project_delete_handler(sender, instance, **kwargs):
-    # Perform actions here
+    print("Project",instance.id)
     print(f'Instance {instance} is being deleted')
 
 @receiver(pre_delete, sender=File)
 def file_delete_handler(sender, instance, **kwargs):
-    # Perform actions here
+    print("File",instance.id)
     print(f'Instance {instance} is being deleted')
 
 @receiver(pre_delete, sender=Job)
 def job_delete_handler(sender, instance, **kwargs):
-    # Perform actions here
+    print("Job",instance.id)
     print(f'Instance {instance} is being deleted')
 
 @receiver(pre_delete, sender=Task)
 def task_delete_handler(sender, instance, **kwargs):
-    # Perform actions here
+    print("Task",instance.id)
     print(f'Instance {instance} is being deleted')
